move files from interfaces bids
