+++ nilearn/_utils/data_gen.py
@@ -15,7 +15,7 @@ from scipy.ndimage import binary_dilation
 
 from nilearn import datasets, image, maskers, masking
 from nilearn._utils import as_ndarray, logger
-from nilearn.interfaces.bids._utils import _bids_entities, _check_bids_label
+from nilearn.interfaces.bids.utils import check_bids_label, bids_entities
 
 
 def generate_mni_space_img(n_scans=1, res=30, random_state=0, mask_dilation=2):
@@ -918,7 +918,7 @@ def create_fake_bids_dataset(
     bids_path = Path(base_dir) / bids_dataset_dir
 
     for task_ in tasks:
-        _check_bids_label(task_)
+        check_bids_label(task_)
 
     if not isinstance(n_runs, list) or not all(
         isinstance(x, int) for x in n_runs
@@ -983,15 +983,15 @@ def _check_entities_and_labels(entities):
         raise ValueError("Only a single extra entity is supported for now.")
 
     for key in entities:
-        if key not in [*_bids_entities()["raw"],
-                       *_bids_entities()["derivatives"]]:
-            allowed_entities = [*_bids_entities()['raw'],
-                                *_bids_entities()['derivatives']]
+        if key not in [*bids_entities()["raw"],
+                       *bids_entities()["derivatives"]]:
+            allowed_entities = [*bids_entities()['raw'],
+                                *bids_entities()['derivatives']]
             raise ValueError(
                 f"Invalid entity: {key}. Allowed entities are: "
                 f"{allowed_entities}"
             )
-        [_check_bids_label(label_) for label_ in entities[key]]
+        [check_bids_label(label_) for label_ in entities[key]]
 
 
 def _mock_bids_dataset(
@@ -1065,7 +1065,7 @@ def _mock_bids_dataset(
                                 task=task,
                                 run=run,
                             )
-                            if key in _bids_entities()["raw"]:
+                            if key in bids_entities()["raw"]:
                                 fields["entities"][key] = label
                             _write_bids_raw_func(
                                 func_path=func_path,
@@ -1223,7 +1223,7 @@ def _create_bids_filename(
 
     """
     if entities_to_include is None:
-        entities_to_include = _bids_entities()["raw"]
+        entities_to_include = bids_entities()["raw"]
 
     filename = ""
 
@@ -1393,7 +1393,7 @@ def _write_bids_derivative_func(
         fields["suffix"] = confounds_tag
         fields["extension"] = "tsv"
         confounds_path = func_path / _create_bids_filename(
-            fields=fields, entities_to_include=_bids_entities()["raw"]
+            fields=fields, entities_to_include=bids_entities()["raw"]
         )
         _basic_confounds(length=n_time_points, random_state=rand_gen).to_csv(
             confounds_path, sep="\t", index=None
@@ -1405,8 +1405,8 @@ def _write_bids_derivative_func(
     shape = [n_voxels, n_voxels, n_voxels, n_time_points]
 
     entities_to_include = [
-        *_bids_entities()["raw"],
-        *_bids_entities()["derivatives"]
+        *bids_entities()["raw"],
+        *bids_entities()["derivatives"]
     ]
 
     for space in ("MNI", "T1w"):
